% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aresimcn.R
\name{aresimcn}
\alias{aresimcn}
\title{A Relative Efficency Investigation Simulation}
\usage{
aresimcn(n = 0, nsims = 1, eps = 0, vc = 1)
}
\arguments{
\item{n}{Size of generated samples}

\item{nsims}{Number of simulations}

\item{eps}{Proportion of contamination}

\item{vc}{Standard deviation of contaminated component}
}
\value{
\code{aresimcn} is an estimate of relative efficency.
}
\description{
Calculates the relative efficency between estimators for finite
samples sizes. The investigation involves making comparisons over families
of distributions, as well as a selection of sample sizes.
}
\details{
An illustration of this function can be found in Section 10.3.4 on
page 596. This example uses samples of size 30 and runs 10,000 simulations.
}
\examples{
aresimcn(30, 1000, 0.25, 3)
aresimcn(30, 100, 0.20, 25)

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
\seealso{
mses() for calculating mean square error and rcn() for generating
random contaminated normal deviates.
}
