% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bb}
\alias{bb}
\title{Baseball}
\format{A list with 59 rows and 6 variables:
\describe{
\item{hand}{right or left}
\item{height}{height in inches}
\item{weight}{weight in pounds}
\item{hitind}{hit index}
\item{hitpitind}{hit pitch index}
\item{average}{batting average}
}}
\usage{
bb
}
\description{
A data set containing data on 59 professional baseball players.
}
\details{
see page 236, 243, 554 of the book.
}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\keyword{datasets}
