% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bday.R
\name{bday}
\alias{bday}
\title{Birthday Solver}
\usage{
bday(n = 0)
}
\arguments{
\item{n}{Sample size of the number of people in a room}
}
\value{
The probability \code{bday} of at least two people having the same
birthdays
}
\description{
Calcuates the probability of at least two out of 'n' people
in a room having the same birthday, assuming that birthdays are equilikely
throughout the year.
}
\details{
Computes the solution to Example 1.3.3 (8th Edition)
}
\examples{
bday(2)
bday(10)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
