% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaplts.R
\name{betaplts}
\alias{betaplts}
\title{Beta Distribution Plots}
\usage{
betaplts()
}
\description{
this function plots the beta distribution for alpha equaling 2,
3, 4, and 5 and beta equaling 2, 3, 4, and 5.
}
\examples{
betaplts()

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
