% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomci.R
\name{binomci}
\alias{binomci}
\title{Binomial distribution Confidence Interval}
\usage{
binomci(s, n, theta1, theta2, value, maxstp = 100, eps = 1e-05)
}
\arguments{
\item{s}{an integer the number of successes}

\item{n}{an integer the the number of trials (zero or more)}

\item{theta1}{a number, the lower bracket probability of success for each trial}

\item{theta2}{a floading point number, the upper bracket probability of success for each trial, must be larger than theta1}

\item{value}{a number, the target distribution function}

\item{maxstp}{an integer default is 100, the amount of times the solution is narrowed down}

\item{eps}{a number default is .00001, the smallest difference in theta1 and theta2}
}
\value{
a list with solution and valatsol (value at solution)

solution a number, the actual confidence interval

valatsol a number, the actual distribution function the solution is found at
}
\description{
This function iteratively solves for the upper and lower
confidence interval bounds for the probaility of success for a binomial sample.
}
\examples{
s <- 17
n <- 30
theta1 <- .4
theta2 <- .45
value <- .95
binomci(s, n, theta1, theta2, value)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
