% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binompower.R
\name{binompower}
\alias{binompower}
\title{Power Functions for Binomial Tests}
\usage{
binompower()
}
\arguments{
\item{n}{Size of sample}

\item{k1}{Size of critical region for Test 1}

\item{k2}{Size of critical region for Test 2}

\item{p0}{Specified power for hypothesis test}
}
\value{
A plot of the power functions for binomial tests of size alpha
}
\description{
Plots the power functions for binomial tests which validate
in general the monotonicity of the power function for binomial tests.
}
\details{
Produces a graph similar to what is seen in Figure 4.5.1 (8th Edition).
The power functions in the plot are described in Example 4.5.2 (8th Edition).
}
\examples{
binompower()

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
