% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapcis64.R
\name{bootstrapcis64}
\alias{bootstrapcis64}
\title{Percentile Bootstrap Confidence Interval}
\usage{
bootstrapcis64(x, c, nb = 3000, alp2 = 0.025)
}
\arguments{
\item{x}{a random variables with sample size \emph{n}}

\item{c}{a number, a range value of intrest}

\item{nb}{a positive integer, the number iteratins to estimate the
confidence interval. defaults to 3000}

\item{alp2}{a number, decides where the lower and upper bound
will be found. defaults to 0.025}
}
\value{
a list containg estimate, lower bound and upper bound

estimate the estimated probability x <= c

lower bound, lower bound of the confidance interval

upper bound, upper bound of the confidence interval
}
\description{
This function returns a bootstrap confidence for the maximum likelihood
estimator that x is less than or eqal to c discussed in Exercise 6.4.7 on page 395.
}
\examples{
x <- c(4, -2, 5.43, 10)
c <- 5
nb <- 3000
alp2 <- 0.025
bootstrapcis64(x, c, nb, alp2)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
