% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottestonemed.R
\name{boottestonemed}
\alias{boottestonemed}
\title{Bootstrap Test One Median}
\usage{
boottestonemed(sample, theta0, b)
}
\arguments{
\item{sample}{a vector, is the sample}

\item{theta0}{a number, the null value of the mean}

\item{b}{an integer, the number of bootstrap resamples}
}
\value{
a list with origtest, pvalue, and teststatall

originaltest is the median of the sample

pvalue is the bootstrap p-value

teststatall contains the bootstrap test median
}
\description{
On page 311 of HMC (2019), a bootstrap test for a one sample
location problem based on the sample mean is presented. For the function, boottestonemed, the sample median is
used for the test statistic. The hypotheses tested, are:H 0 :
\eqn{\Theta = \Theta 0} versus H a : \eqn{\Theta > \Theta 0} where
\eqn{\Theta 0} is spectified, (an input value of the function).
The bootstrapped p-value is returned.
}
\examples{
sample <- c(4, 3, 2, 9)
theta0 <- .05
b <- 5
boottestonemed(sample, theta0, b)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
