% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chistable.R
\name{chistable}
\alias{chistable}
\title{Chi-Square Distribution}
\usage{
chistable()
}
\value{
\code{chistable} is the chi-square distribution table for the selected
quantiles.
}
\description{
Presents selected quantiles of the chi-square distribution.
}
\details{
This table more specifically produces the chi-square distribution table
for the \emph{0.01}, \emph{0.025}, \emph{0.050}, \emph{0.100}, \emph{0.900}, \emph{0.950},
\emph{0.975}, and \emph{0.990} quantiles. The initial row indicates the selected
quantiles and the inital column indicates 'r' degrees of freedom See page 708
for more details.
}
\examples{
chistable()

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
