% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cipi.R
\name{cipi}
\alias{cipi}
\title{Confidence Interval and Predictive Interval for Linear Models}
\usage{
cipi(fit, hmat, alpha = 0.05)
}
\arguments{
\item{fit}{A linear model}

\item{hmat}{A vector used to store the value to be predicted for}

\item{alpha}{Level of significance to calculate the confidence interval
and predictive interval.}
}
\value{
n x 1 numeric vector containing the random deviates.
}
\description{
Calculates the predicted response value associated with the
inputted hmat value. This function also calculates the predicted value's
confidence interval and predictive interval along with their standard errors.
}
\details{
A sample of this function's usage can be found in Example 9.6.1
from the book, Introduction to Mathematical Statistics (8th Edition)
}
\examples{
year <- men1500m$year
time <- men1500m$time
cipi(fit = lm(time ~ year), hmat = c(1, 2020), alpha = 0.05)

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
