% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condsim1.r
\name{condsim1}
\alias{condsim1}
\title{Observation Generator}
\usage{
condsim1(nsims)
}
\arguments{
\item{nsims}{a positive non-zero number, the amout of times you want it
to generate an observation}
}
\value{
a list of numeric observations
}
\description{
this algorithm generates observations from the probability
density function
}
\details{
observation is generated with this Mathmatical function:
log(1-runif(1))+.5*log(1-runif(1))
}
\examples{
nsims <- 5
obs <- condsim1(nsims)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
