% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{crimealk}
\alias{crimealk}
\title{crimealk}
\format{A matrix containg 6 crimes and how many drinkers vs non drinkers
commited them
\describe{
\item{arson}{a row corresponding to people who commited arson}
\item{rape}{a row corresponding to people who commited rape}
\item{violence}{a row corresponding to people who commited violent crimes}
\item{theft}{a row corresponding to people who commited theft}
\item{coining}{a row corresponding to people who commited coining}
\item{fraud}{a row corresponding to people who commited fraud}
\item{drinker}{a column corresponding to the people who drink alcohol}
\item{nondrinker}{a column corresponding to the people who do not drink alcohol}
}}
\usage{
crimealk
}
\description{
A data set containing the number of different crimes commited by drinkers
and non drinkers
}
\details{
see page 291 of the book.
}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\keyword{datasets}
