% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{darwin}
\alias{darwin}
\title{darwin}
\format{A list containing 15 rows with 2 variables
\describe{
  \item{cross}{no clue}
  \item{self}{no cluer}
}}
\usage{
darwin
}
\description{
A data set containing 15 observation from darwin
}
\keyword{datasets}
