% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dex165.R
\name{dex165}
\alias{dex165}
\title{probability mass function for exersize 1.6.5}
\usage{
dex165()
}
\value{
a matrix with the value of x and its pmf

x the x value used to determine the output for the pmf

pmf the output of the probability mass function using the corresponding x value
}
\description{
a pmf for choosing 5 out of 100 units to test for acceptability
}
\details{
pmf for Exercise 1.6.5 part a
}
\examples{
pmf_matrix <- dex165()

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
