% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlogF.R
\name{dlogF}
\alias{dlogF}
\title{Density of Log Distribution}
\usage{
dlogF(x)
}
\arguments{
\item{x}{Real number (or vector of numbers) where -inf < x < inf.
In R, these limits are determined by data limitations.
Thus we get the following piecewise function,\preformatted{           \{   0,          x <= -746
dlogF(x) = \{ pdf,  -745 <= x <=  589
           \{   0,   590 <= x <=  709
           \{ NaN,          x >=  710
}}
}
\value{
PDF (single value or vector based on input) of X.
}
\description{
Computes the probability density function (PDF) of a log F random variable.
The PDF is given in expression 1.7.9 on page 52 of the book.
}
\examples{
pdf <- dlogF(0)
pdf <- dlogF(589)
pdf <- dlogF(3.14159)

# Example where input is a vector of numbers to be passed into function.
# Each number in this input vector is evaluated independently.
v <- c(1, 2, 3, 4, 5, 10, 20.5)
pdf_vec <- dlogF(v)

}
\references{
Hogg, R. McKean, J. Craig, A (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson
}
