% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloggamma.R
\name{dloggamma}
\alias{dloggamma}
\title{Density of Log Gamma Distribution}
\usage{
dloggamma(x, alpha, beta)
}
\arguments{
\item{x}{a numeric Vector. All components are greater than 1.}

\item{alpha}{a positive non-zero number corresponds to the shape.}

\item{beta}{a positive non-zero number corresponds to the rate.}
}
\value{
The density of the log gamma distribution
}
\description{
this function calculates density of log gamma distribution.
}
\details{
f(x) = (1/(gamma(alpha)*beta^alpha))*x^(-(1+beta)/beta)*(log x)^(alpha-1)
}
\examples{
dloggamma(5, 1, 1)

}
\references{
Hogg, R. McKean, J. Craig, A (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson
}
