% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{eg4.4.4data}
\alias{eg4.4.4data}
\title{Eg 4.4.4 Data}
\format{A list with 15 rows and 1 variable
\describe{
\item{x}{some variable from 4.4.4}
}}
\usage{
eg4.4.4data
}
\description{
A data set containing 15 values.
}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\keyword{datasets}
