% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{elasticmod}
\alias{elasticmod}
\title{elasticmod}
\format{A list with 22 rows and 2 variables
\describe{
  \item{elasticmod}{not sure}
  \item{ind}{dont know}
}}
\usage{
elasticmod
}
\description{
A data set containg 22 observations related to elastic
}
\keyword{datasets}
