% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipmake.R
\name{ellipmake}
\alias{ellipmake}
\title{Plot Ellipse}
\usage{
ellipmake(p = 0.95, b = matrix(c(1, 0.75, 0.75, 1), nrow = 2),
  mu = c(5, 2))
}
\arguments{
\item{p}{floating point number, between 0 and 1 adjusts the size of the ellipse}

\item{b}{a 2x2 matrix whith a determinate grater than 0 and all elements are positive, adjusts the shape of the ellipse}

\item{mu}{a vector of length 2, adjusts the center of the ellipse}
}
\description{
this function plots an ellipse of varying size and coridnates of
the center
}
\examples{
p <- .85
b <- matrix(c(2,1,1,3), nrow=2)
mu <- c(3, 3)
ellipmake(p, b, mu)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
