% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empalphacn.R
\name{empalphacn}
\alias{empalphacn}
\title{emperical alpha}
\usage{
empalphacn(nsims)
}
\arguments{
\item{nsims}{an integer, the number of simulations that are to be run}
}
\value{
a list with empiricalalpha and error.

empiricalalpha is a numeric value that is the empirical alpha.

error is a neumeric value that is the calculated error
}
\description{
Obtains the empirical level of the probability
density function where: f(x) = 4x^3 when 0 < x < 1; and 0 otherwise.
}
\examples{
nsims <- 10
empalpha_error <- empalphacn(nsims)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\seealso{
\code{\link{rcn}} for  generating random contaminated normal deviates.
}
