% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex1427.R
\name{ex1427}
\alias{ex1427}
\title{Number drawing game}
\usage{
ex1427()
}
\value{
returns 1 if the player wins; else returns 0
}
\description{
in this game first a number is drawn from a set 1-20, let x
be the number drawn. Next another number is drawn from a set x-25 if the
second number is greater than 21 then its a win.
}
\details{
game simulation for Exercise 1.4.27 on page 36
}
\examples{
result <- ex1427()

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
