% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex158.R
\name{ex158}
\alias{ex158}
\title{piecewise cdf graph}
\usage{
ex158()
}
\description{
creates a graph where F(x)=0 when x<-1,
F(x)=(x+2)/4 when -1<=x<1, and F(x)=1 where 1<=x.
}
\details{
creates the cdf for Exercise 1.5.8.
}
\examples{
ex158()

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
