% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex4.4.3data}
\alias{ex4.4.3data}
\title{ex 4.4.3 data}
\format{A list containing 21 rows and 1 variable
\describe{
  \item{x}{some variable from 4.4.3}
}}
\usage{
ex4.4.3data
}
\description{
A data set containg 21 values
}
\keyword{datasets}
