% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex6111}
\alias{ex6111}
\title{ex 6111 data}
\format{A list containing 25 rows and 1 variable
\describe{
  \item{x}{some variable from 6111}
}}
\usage{
ex6111
}
\description{
A data set containg 25 values
}
\keyword{datasets}
