% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{examp1053}
\alias{examp1053}
\title{ex 1053 data}
\format{A list containing 15 rows and 2 variables
\describe{
  \item{x}{some variable from 1053}
  \item{y}{some variable frim 1053}
}}
\usage{
examp1053
}
\description{
A data set containg 15 x and y values
}
\keyword{datasets}
