% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fastcars}
\alias{fastcars}
\title{Fast Cars}
\format{A list containing 30 rows and 2 variables
\describe{
\item{ind}{type of car}
\item{speed}{the speed of the car}
}}
\usage{
fastcars
}
\description{
A data set containg 30 observations of car speeds.
}
\details{
see page 528 of the book.
}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\keyword{datasets}
