% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp1.R
\name{fp1}
\alias{fp1}
\title{F-Distribution Table for Upper 0.05 Critical Points}
\usage{
fp1()
}
\value{
\code{xmat} is the F-distribution table for the 95th quantile.
}
\description{
Produces the F-distribution table for the 0.95 quantile
of the F-distribution.
}
\details{
This table more specifically produces the F-distribution table for
when alpha equals 0.05. The initial row indicates the first nine degrees
of freedom in the numerator and the inital column indicates degrees of freedom in
the denominator. See page 711 for more details.
}
\examples{
fp1()

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
\seealso{
fp2() for the F-distribution table of the 0.95 quantile with the numerator
having greater than nine degrees of freedom., as well as fp3() and
fp4() for the F-distribution table of the 0.99 quantile.
}
