% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp3.R
\name{fp3}
\alias{fp3}
\title{F-Distribution Table for Upper 0.01 Critical Points}
\usage{
fp3()
}
\value{
\code{xmat} is the F-distribution table for the 99th quantile
}
\description{
Produces the F-distribution table for the 0.99 quantile
of the F-distribution.
}
\details{
This table more specifically produces the F-distribution table for
when alpha equals 0.01. The initial row indicates the first nine degrees
of freedom in the numerator and the inital column indicates degrees of freedom in
the denominator. See page 713 for more details.
}
\examples{
fp3()

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
\seealso{
fp4() for the F-distribution table of the 0.99 quantilss with
the numerator having greater than nine degrees of freedom, as well as fp1() and
fp2() for the F-distribution table of the 0.95 quantile.
}
