% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftableb.R
\name{ftableb}
\alias{ftableb}
\title{F-Distribution Table for 0.05, 0.25, and 0.01 Critical Points}
\usage{
ftableb()
}
\value{
\code{xmat} is the F-distribution table for the 95th quantile.
}
\description{
Produces the F-distribution table for the 0.950, 0.975,
and 0.990 quantiles of the F-distribution. The first column of the table
consists of these three quantiles repeating in this fashion. The second column
represents degrees of freedom in the denominator for each of these critical points.
The rest of the columns are for the first 11 degrees of freedom in the numerator.
}
\examples{
ftableb()

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
            Statistics, 8th Ed. Boston: Pearson.
}
\seealso{
fp1() for the F-distribution table of the 0.95 quantile with
the numerator having less than ten degrees of freedom, fp2() for the
F-distribution table of the 0.95 quantile with the numerator,
having greater than nine degrees of freedom., as well as fp3() and
fp4() for the F-distribution table of the 0.99 quantile.
}
