% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{genexpd}
\alias{genexpd}
\title{genexpd}
\format{A list containing 22 rows and 2 variables
\describe{
  \item{c.x..y.}{i dont know}
  \item{indy}{not sure}
}}
\usage{
genexpd
}
\description{
A data set containg 22 observations about genexpd
}
\keyword{datasets}
