% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcis.R
\name{getcis}
\alias{getcis}
\title{Get Confidence Intervals}
\usage{
getcis(mat, cc = 0.9)
}
\arguments{
\item{mat}{an m x n matrix}

\item{cc}{confidence coefficient, or the confidence level of the intervals}
}
\value{
(1 - alpha)100\% confidence intervals in a m x 2 matrix
}
\description{
Calculates t confidence intervals for m samples each of size n
contained in the R matrix mat which is a \emph{m x n} matrix. Each row is a
sample of size n. The function is discussed in Exercise 4.2.11 on page 245 of
HMC (2018).
}
\details{
This function is used to calculate t confidence intervals for m
samples each of size n contained in the R matrix mat, which is a \emph{m x n}
matrix. Each row is a sample of size n.
}
\examples{
n <- 10
m <- 50
mat <- matrix(rnorm(m * n), ncol = n)
getcis(mat = mat, cc = 0.95)
getcis(mat = mat, cc = 0.90)

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
            Statistics, 8th Ed. Boston: Pearson.
}
