% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbser3.R
\name{gibbser3}
\alias{gibbser3}
\title{Gibbs Sampler with Bayesian Inference}
\usage{
gibbser3(alpha, beta = 0, nt = 0, m = 0, n = 1)
}
\arguments{
\item{alpha}{test statistic applied in algorithm}

\item{beta}{NEED INFO}

\item{nt}{NEED INFO}

\item{m}{length of returned variables x1 and y1}

\item{n}{length of returned variables x2 and y2}
}
\value{
A list of two streams of iid random variables. The first pair is of
length \emph{m} and the second of length \emph{n}.
}
\description{
This function computes the Gibbs sampler routine as discussed on
page 677, but with Bayesian influence integrated. The input parameter alpha
is the parameter in the joint distribution of the random vector (X, Y) which
is given in the example. The value of m is the number of initial runs of the
sampler for achieving equilibrium; then the next (n * m) runs are recorded and returned
in the R vectors x2 and y2. The paired items (x2, i) & (y2, i) are the variates for
random vector (X, Y).
}
\examples{
gibbser3(alpha = 10, beta = 4, nt = 100, m = 30, n = 60)
gibbser3(alpha = 12, beta = 4, nt = 100, m = 40, n = 100)

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
\seealso{
gibbser2() for more details on the Gibbs sampler routine described on
page 677.
}
