% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarch1.R
\name{hierarch1}
\alias{hierarch1}
\title{Hierarchal Bayes Model}
\usage{
hierarch1(nsims = 0, x = 0, tau = 0.05, kstart = 1)
}
\arguments{
\item{nsims}{Number of iterations for the Gibbs sampler}

\item{x}{Value observed}

\item{tau}{Statistic used to measure the ordinal association between}

\item{kstart}{Indicates at which value of nsims the Gibbs sample commences}
}
\value{
**NEED INFO
}
\description{
**NEED INFO
}
\details{
Computes the solution to Example 11.4.2 on page 681.
}
\examples{
hierarch1(nsims = 300, x = 6, tau = 0.05, kstart = 20)
hierarch1(nsims = 150, x = 3, tau = 0.045, kstart = 35)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
\seealso{
\code{\link{gibbser2}}
}
