% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lengthriver}
\alias{lengthriver}
\title{river length}
\format{A list containing 141 rows and 1 variable
\describe{
  \item{lengthriver}{the length of rivers}
}}
\usage{
lengthriver
}
\description{
A data set containg 141 observations of values of river lenght
}
\keyword{datasets}
