% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lifetimemotor}
\alias{lifetimemotor}
\title{Life Time of Motors}
\format{A list containing 20 rows and 1 variable
\describe{
\item{lifetime}{life time of a motor in hours}
}}
\usage{
lifetimemotor
}
\description{
A data set containg 20 observations about motor life time in high
tempatures.
}
\details{
see page 235 of the book.
}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\keyword{datasets}
