% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpbon.R
\name{mcpbon}
\alias{mcpbon}
\title{Bonferroni's Multiple Comparisons Procedure}
\usage{
mcpbon(y, ind, alpha = 0.05)
}
\arguments{
\item{y}{Vector of the combined samples of size n}

\item{ind}{Corresponding treatment vector of size n}

\item{alpha}{Level of significance used for Bonferroni Procedure}
}
\value{
This function returns a table to the user. The first two columns
\code{j} \& \code{jp} are the specific levels (categories of cars) being compared.
\code{muj} \& \code{mujp} are the means of two levels being compared. \code{diff}
is the difference in means between these levels. \code{se} is the standard
error for each comparison, \code{err} is the comparison's error estimate, and
\code{lb} \& \code{ub} consist of the lower and upper bound for each comparison's
confidence interval.
}
\description{
Bonferroni's approach to Multiple Comparisons Procedure that uses
the method of complements, DeMorgan's Laws and Boole's inequality. This procedure
is robust and can be used in many settings rather than just one-way design.
}
\details{
Computes the Bonferroni confidence intervals as seen in Example
9.4.1 on page 529.
}
\examples{
load('./data/fastcars.rda')
speed <- fastcars$speed
car <- fastcars$ind
mcpbon(speed, car)

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
            Statistics, 8th Ed. Boston: Pearson.
}
\seealso{
mcpfisher() for details about Fisher's PLSD Multiple Comparison Procedure
}
