% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnormal.R
\name{mixnormal}
\alias{mixnormal}
\title{The First Iteration of the EM Step}
\usage{
mixnormal(x, theta0)
}
\arguments{
\item{x}{a vector of quantiles.}

\item{theta0}{a vector, first 2 elements are means, next 2
elements are standard deviations, and the last element is
the perportion of data from the first population}
}
\value{
returns a vector, with mu1, mu2, sig1, sig2 and p
}
\description{
This function returns one iteration of the EM
step for excersize 6.6.8. The initial estimate for the step
is theinput vector theta0 while the ouput are the onestep
esimates as given on page 409 of the text.
}
\examples{
x <- c(5, 7, 2, 1,9)
theta0 <- c(.2, .6, .8, .9, .95)
mixnormal(x, theta0)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
