% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mses.R
\name{mses}
\alias{mses}
\title{Mean Squared Error}
\usage{
mses(x, theta0 = 0)
}
\arguments{
\item{x}{A numeric vector containing sample size \emph{n}}

\item{theta0}{Variance of sample \emph{x}}
}
\value{
The mean squared error of the inputted sample and value for theta.
}
\description{
As part of the Monte Carlo investigation of the relative efficency
between estimators for finite sample sizes, this function calculates the sum
of squared differences of the inputted sample observations against the
value of \code{theta0}. The sum is divided by size of the sample \emph{n}.
}
\details{
The Monte Carlo investigation of relative efficency between estimators
for finite sample sizes is covered in Section 10.3.4 on page 595.
}
\examples{
sample1 <- rcn(n = 30, eps = 0.20, sigma_c = 25)
sample2 <- rcn(n = 20, eps = 0.20, sigma_c = 25)

theta1 <- var(sample1)
theta2 <- var(sample2)

MSE1 <- mses(sample1, theta1)
MSE2 <- mses(sample2, theta2)

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
