% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrial.R
\name{multitrial}
\alias{multitrial}
\title{Multinomial Trial Simulation}
\usage{
multitrial(p)
}
\arguments{
\item{p}{A vector containing the probability of outcome for k levels of a random
multinomial trial}
}
\value{
\code{multitrial} contains the outcome of one event for a random multinomial trial.
}
\description{
Computes the result of a random multinomial trial based on the inputted
probability of outcome for each k level.
}
\details{
An illustration of this function's use can be found in Exercise 3.1.19
on page 165.
}
\examples{
ps <- c(0.3, 0.2, 0.2, 0.2, 0.1)

# Performing 10 random trials if ps <- c(0.3, 0.2, 0.2, 0.2, 0.1)
trials <- 1:10
result <- rep(0, length(trials))

for (i in 1:length(trials)) {
    result[i] <- multitrial(ps)
}

# Result of 10 random trials
result

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
            Statistics, 8th Ed. Boston: Pearson.
}
\seealso{
psum() for ...
}
