% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfigc3s3.1.R
\name{newfigc3s3.1}
\alias{newfigc3s3.1}
\title{Figure 3.3.1}
\usage{
newfigc3s3.1()
}
\value{
Two gamma distribution plots with various inputs for paramaters
\emph{alpha} and \emph{beta}.
}
\description{
This function produces several gamma distributions with various
inputs for the parameters \emph{alpha} and \emph{beta}. The top plot illustrates
how changing values for \emph{alpha} influences the gamma distribution with
\emph{beta} set to equal 4. The plot below illustrates how changing values for
\emph{beta} influences the gamma distribution with \emph{alpha} set to to equal
4.
}
\details{
The images and descriptions of these graphs can also be found on pages
175 and 176 in the book.
}
\examples{
newfigc3s3.1()

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
