% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{normal50}
\alias{normal50}
\title{normal50}
\format{A list containing 50 rows and 1 variable
\describe{
  \item{x}{some x value}
}}
\usage{
normal50
}
\description{
A data set containg 50 observations of x values
}
\keyword{datasets}
