% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaltable.R
\name{normaltable}
\alias{normaltable}
\title{Normal Distribution Table}
\usage{
normaltable()
}
\value{
\code{pz} is the normal distribution table for values of \emph{z} that is
greater than or equal to zero.
}
\description{
Generates the normal probabilities for values of \emph{z} that is
greater than or equal to zero. See page 709 for more details.
}
\examples{
normaltable()

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
