% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{olym1500mara}
\alias{olym1500mara}
\title{olympic 1500 marathon}
\format{A list containing 20 rows and 3 variables
\describe{
  \item{year}{the year it took place}
  \item{m1500}{not sure}
  \item{marathon}{not sure}
}}
\usage{
olym1500mara
}
\description{
A data set containg 20 observations about olympic 1500 marathon
}
\keyword{datasets}
