% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesampsgn.R
\name{onesampsgn}
\alias{onesampsgn}
\title{One-Sample Sign Analysis}
\usage{
onesampsgn(x, test = FALSE, alt = 0, theta0 = 0, alpha = 0.05,
  maktable = TRUE, plotb = FALSE)
}
\arguments{
\item{x}{Sample of size 'n'}

\item{test}{Need Info}

\item{alt}{Need Info}

\item{theta0}{Variance of sample \emph{x}}

\item{alpha}{Level of significance}

\item{maktable}{If true provides more descriptive output for x}

\item{plotb}{If true a boxplot is produced}
}
\value{
The probability \code{bday} of at least two people having the same
birthdays
}
\description{
Calculates a distribution free confidence interval for a sample
median by deriving for the 100pth distribution percentile of a continuous random
variable.
.
}
\details{
Computes the solution to Example 1.3.3 (8th Edition)
}
\examples{
x <- c(56, 70, 89, 94, 96, 101, 102, 102, 102, 105, 106, 108, 110, 113, 116)
# Confidence interval for the median:
onesampsgn(x, alpha = 0.12)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearson.
}
