% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentciboot.R
\name{percentciboot}
\alias{percentciboot}
\title{Percent Confidence Interval bootstrap}
\usage{
percentciboot(sample, b, alpha = 0.05)
}
\arguments{
\item{sample, }{the original sample.}

\item{b}{an integer the desired numbers of bootstraps.}

\item{alpha}{a number 1 - alpha is the confidence coefficient.}
}
\value{
a list with thata, lower, upper, thetastar

thata is the point estimate.

lower is the lower end of the percentile confidence interval.

upper is the upper end of the percentile confidence interval.

thetastar is the vector of bootstrapped thetastars
}
\description{
obtains the percentile confidence interval for the mean.
}
\examples{
sample <- c(4, 3, 2, 3)
b <- 20
alpha <- .05
percentciboot(sample, b, alpha)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
