% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piest.r
\name{piest}
\alias{piest}
\title{Estimation of Pi Using Binomial Experiment}
\usage{
piest(n)
}
\arguments{
\item{n}{Iterations (number of simulations) to repeat.}
}
\value{
Tuple containing estimate value of pi and standard error.
}
\description{
Uses iterative methods to estimate pi with standard error for
the simulation discussed in Example 4.8.1 on page 293 of the book.
}
\examples{
# Example where parameters are passed in as variables.
n <- 1000
result <- piest(n)

# Example where parameters are passed in as values.
result <- piest(1000)

}
\references{
Hogg, R. McKean, J. Craig, A (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson
}
