% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piest2.R
\name{piest2}
\alias{piest2}
\title{Estimation of Pi by Monte Carlo Integration}
\usage{
piest2(n)
}
\arguments{
\item{n}{Sample size.}
}
\value{
Tuple containing estimate value of pi and the standard error.
}
\description{
Uses integration to estimate pi given a sample size.
See example 4.8.4 on page 295 and 296 of the book.
}
\examples{
# Example where parameters are passed in as variables.
n <- 1000
result <- piest2(n)

# Example where parameters are passed in as values.
result <- piest2(1000)

}
\references{
Hogg, R. McKean, J. Craig, A (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson
}
