% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plogF.r
\name{plogF}
\alias{plogF}
\title{Probability of Log Distribution}
\usage{
plogF(x)
}
\arguments{
\item{x}{Real number (or vector of numbers) where \eqn{-\infty < x < \infty}.
CDF of X is \eqn{F(x) = 1 - (1+5e^(x))^(-0.2)}.}
}
\value{
CDF (single value or vector based on input) of X. \eqn{P(X\le x)}
}
\description{
Computes the distribution function (CDF) of a log F random variable.
The PDF is given in expression 1.7.9 on page 52 of the book.
}
\examples{
cdf <- plogF(1)
cdf <- plogF(-1)
cdf <- plogF(3.14159)
}
\references{
Hogg, R. McKean, J. Craig, A (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson
}
