% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploggamma.r
\name{ploggamma}
\alias{ploggamma}
\title{PDF of Mixture Distribution}
\usage{
ploggamma(x, alpha, beta)
}
\arguments{
\item{x}{Vector of quantiles used in pgamma() function. All components are >= 1.}

\item{alpha}{Shape value used in pgamma() function. Must be positive.}

\item{beta}{Rate value used in pgamma() function. Cannot be zero.}
}
\value{
Distribution function from pgamma() function.
}
\description{
Computes the PDF of the loggamma of the continuous-type random variable X given the following:
P(a < x < b) = definite integral from z to b of (f(x)dx)
f(x) is the density
See Remark 3.7.1 on page 219 of the book.
}
\examples{
# Example where parameters are passed in as variables.
x <- 10
alpha <- 1
beta <- 1
result <- ploggamma(x, alpha, beta)

# Example where parameters are passed in as values.
result <- ploggamma(10, 1, 1)

}
\references{
Hogg, R. McKean, J. Craig, A (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson
}
