% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsulfur.r
\name{plotsulfur}
\alias{plotsulfur}
\title{plot sulfurdioxide}
\usage{
plotsulfur()
}
\description{
This function creates a histogram using the sulfurdioxide table
from the sulfrurdio.rda.
}
\examples{
plotsulfur()

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
