% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltlogF.r
\name{pltlogF}
\alias{pltlogF}
\title{Plot Log Function}
\usage{
pltlogF()
}
\description{
this function crates a file and plots the log function to that file
}
\examples{
pltlogF()

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
