% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisrand.R
\name{poisrand}
\alias{poisrand}
\title{Poisson Random Sample Generator}
\usage{
poisrand(n, lambda)
}
\arguments{
\item{n}{a positive non-zero integer the, size of sample}

\item{lambda}{a positive non-zero floating point number, The mean of the poisson distribution}
}
\value{
a poisson random sample
}
\description{
This function creates a random sample of n elements using the
poisson distribution.
}
\details{
Simulates the Poisson Processes in Example 4.8.2 on page 295 of HMC
(2018).
}
\examples{
n <- 10
lambda <- 25.2
poisrand(n, lambda)
mean(poisrand(10000, lambda))

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to Mathematical
Statistics, 8th Ed. Boston: Pearlson
}
