% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonci.r
\name{poissonci}
\alias{poissonci}
\title{Confidence Interval for mean of of a Poisson Distribution}
\usage{
poissonci(s, n, theta1, theta2, value, maxstp = 100, eps = 1e-05)
}
\arguments{
\item{s}{a number, the number of events that have occured calculted as n * xbar}

\item{n}{a number, the size}

\item{theta1}{a number n*theta1 is a mean and the lower bracket for the solution, must be positive.}

\item{theta2}{a number n*thata2 is a mean and the upper bracket for the solution, must be larger than theta1}

\item{value}{a number the target distribution function}

\item{maxstp}{an integer default is 100, the amount of times the solution is narrowed down}

\item{eps}{a number default is .00001, the smallest difference in theta1 and theta2 as they are updated}
}
\value{
a list with solution and valatsol (value at solution)

solution a floating point number, the actual confidence interval

valatsol a floating point number, the actual distribution function solution is found at
}
\description{
This function iteratively solves for the upper and lower
confidence interval bounds for the probability of success for a poisson
sample.
}
\examples{
s <- 125
n <- 25
theta1 <- 5.5
theta2 <- 6
value <- .05
poissonci(s, n, theta1, theta2, value)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
