% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psum.r
\name{psum}
\alias{psum}
\title{Probability Outcomes for a Multinomial Trial}
\usage{
psum(ps)
}
\arguments{
\item{ps}{k by 1 vector containing probabilities for k possible outcomes}
}
\value{
\code{psum} contains k parameters of a multinomial distribution
}
\description{
\code{psum} is an auxiliary function used in the function \code{multitrial}
which generates a random trial from a multinormal distribution. If there are k categories
for a multinomial with respective probabilities in the k by 1 vector ps then
psum(ps) returns the accumulating probabilities {p_1, p_1+p_2, ..., 1}.
}
\details{
An illustration of this function's use can be found in Exercise 3.1.19
on page 165.
}
\examples{
ps <- c(0.1, 0.1, 0.2, 0.3, 0.2, 0.1)
psum(ps)

ps <- c(0.2, 0.2, 0.2, 0.2, 0.2)
psum(ps)
}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
            Statistics, 8th Ed. Boston: Pearson.
}
\seealso{
multitrial() for more on computing multinomial trials
}
