% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{punter}
\alias{punter}
\title{punter}
\format{A list containing 13 rows and 7 variables
\describe{
  \item{distance}{the distance in some unit}
  \item{strength}{the strength in some unit}
  \item{hangtime}{the time the ball was in the air}
  \item{lf}{no idea}
  \item{llstrength}{no idea}
  \item{rf}{no idea}
  \item{rl}{no idea}
}}
\usage{
punter
}
\description{
A data set containg 13 observations about punters
}
\keyword{datasets}
