% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlogF.r
\name{qlogF}
\alias{qlogF}
\title{Quantile of Log Distribution}
\usage{
qlogF(x)
}
\arguments{
\item{x}{Real number(or vector of numbers) where 0 < x < 1.
CDF of X is F(x) = 1 - (1+5e^(-x))^(-0.1), -inf < x < inf.
# X is a real number with the following piecewise function:
```
           \{  NaN,      x < 0
           \{ -Inf,      x = 0
qlogF(x) = \{    n,  0 < x < 1
           \{  Inf,      x = 1
           \{  NaN,      x > 1
```}
}
\value{
Inverse of the CDF (single value or vector based on input) of X.
}
\description{
Computes the quantiles of the distribution of a log F random variable.
The PDF is given in expression 1.7.9 on page 52 of the book.
}
\examples{
inverse_cdf <- qlogF(0.1)
inverse_cdf <- qlogF(0.9)

}
\references{
Hogg, R. McKean, J. Craig, A (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson
}
