% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplotc4s2.r
\name{qqplotc4s2}
\alias{qqplotc4s2}
\title{Population Quantile Plots}
\usage{
qqplotc4s2(vec)
}
\arguments{
\item{vec}{Vector containing data for graph.}
}
\description{
Function which obtains the normal, Laplace, and
exponential quantiles as shown in Figure 4.4.1.
See Example 4.4.6 on pages 260 and 261 of the book.
}
\examples{
# Example where parameters are passed in as variables.
vec <- 1:1000
qqplotc4s2(vec)

# Example where parameters are passed in as values.
qqplotc4s2(1:1000)

}
\references{
Hogg, R. McKean, J. Craig, A (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson
}
