% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcn.r
\name{rcn}
\alias{rcn}
\title{Random Contaminated Normal Deviates}
\usage{
rcn(n = 0, eps = 0, sigma_c = 1)
}
\arguments{
\item{n}{Size of random sample to generate}

\item{eps}{Proportion of contamination}

\item{sigma_c}{Standard deviation of contaminated component}
}
\value{
n x 1 numeric vector containing the random deviates.
}
\description{
Generates data sets from a contaminated normal distribution as
              in Example 4.9.2
}
\details{
With probability (1-eps) a deviates are drawn from a standard normal
         distribution. With probability eps deviates are drawn from a normal
         distribution with mean 0 and standard devation sigma_c.
}
\examples{
rcn(n = 12, eps = 0.20, sigma_c = 4)
rcn(n = 30, eps = 0.30, sigma_c = 6)

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
            Statistics, 8th Ed. Boston: Pearson.
}
