% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{regr1}
\alias{regr1}
\title{ACT and Exam Scores}
\format{A list containing 15 rows and 2 variables
\describe{
\item{actscores}{score on act}
\item{examscores}{score on exam, out of 200}
}}
\usage{
regr1
}
\description{
A data set containg 15 observations about test scores on ACT and
the final exam in the first semester of calculus.
}
\details{
See page 548 of the book.
}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\keyword{datasets}
