% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscn.r
\name{rscn}
\alias{rscn}
\title{random sample from contaminatied normal distribution}
\usage{
rscn(n, eps, sd, mu)
}
\arguments{
\item{n}{an integer, the size of the sample drawn from the distribution}

\item{eps}{a number, epsilon usually between 0 and 1}

\item{sd}{a number, the standard deviation should be greater than or equal to 1}

\item{mu}{a number, the mean}
}
\value{
a random sample from contaminated normal distribution
}
\description{
simulates a random sample of size n drawn from a skewed contaminated normal
distribution with percent contamination eps, standard deviation ratio
sigma_c and mean mu_c.
}
\examples{
n <- 5
eps <- .2
sd <- 2.4
mu <- 0
rscn(n, eps, sd, mu)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
