% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sec95set2}
\alias{sec95set2}
\title{sec95set2}
\format{A list containing 15 rows and 3 variables
\describe{
\item{x}{some variable x}
\item{fa}{some other variable fa}
\item{fb}{soem other variable fb}
}}
\usage{
sec95set2
}
\description{
A data set containg 15 observations about sec 9.5 set 2
}
\details{
See page 539 of the book.
}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\keyword{datasets}
