% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sect76data}
\alias{sect76data}
\title{section 7.6 data}
\format{A list containing 20 rows and 1 variable
\describe{
  \item{x}{some x value}
}}
\usage{
sect76data
}
\description{
A data set containg 20 observations of x values
}
\keyword{datasets}
