% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{selfrival}
\alias{selfrival}
\title{self vs rival}
\format{A list containing 50 rows and 2 variables
\describe{
  \item{self}{what you got}
  \item{rival}{what your rival got}
}}
\usage{
selfrival
}
\description{
A data set containg 20 observations of self and rival
}
\keyword{datasets}
