% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplegame.r
\name{simplegame}
\alias{simplegame}
\title{Simple Game for Estimation of Expected Gain}
\usage{
simplegame(amtpaid)
}
\arguments{
\item{amtpaid}{Initial amount that a player pays (p_0 in description).}
}
\value{
Estimation of expected gain for player.
}
\description{
Computes the estimation of expected gain for a game wherein the following rules apply:
A player pays p_0 to play. The player then rolls a 6-sided die. If the roll is a 1 or a 2 then the game is over.
Otherwise, the player flips a fair coin. If the coin toss results in a tail, they receive $1 and the game is over.
Otherwise they draw 2 cards without replacement from the standard deck of 52 cards. If none of the cards is an ace
they receive $2, while they receive $10 or $50 is they get 1 or 2 aces respectively. In both cases, the game is over.
Let G denote the player's gain. To determine the expected gain, we need the distribution of G. The support of G is
the set {-p_0, 1-p_0, 2-p_0, 10-p_0, 50-p_0}. For the associated probabilities we need the distribution of X where
X is the number of aces in a draw of 2 cards. The distribution is
\eqn{P(X=x) = ( (4 choose x) * (48 choose (2-x)) ) / (52 choose 2) for x=0,1,2}

See example 1.8.9 on page 65 of the book.
}
\examples{
player_winnings <- simplegame(1)
player_winnings <- simplegame(5)
player_winnings <- simplegame(10)

}
\references{
Hogg, R. McKean, J. Craig, A (2018) Introduction to Mathematical Statistics, 8th Ed. Boston: Pearson
}
