% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{speedlight}
\alias{speedlight}
\title{Speed of Light}
\format{A list containing 66 rows and 1 variable
\describe{
  \item{speed}{the speed of light}
}}
\usage{
speedlight
}
\description{
A data set containg 66 observations of the speed of light
by the scientist Simon Newcomb
}
\details{
See page 238 of the book.
}
\references{
obtained ath Carnegie Melon site http://lib.stat.cmu.edu/DASL/
}
\keyword{datasets}
