% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpower2g2.r
\name{tpower2g2}
\alias{tpower2g2}
\title{Power Function Of The Two Sample Test}
\usage{
tpower2g2(n, m, Delta, sig, alpha)
}
\arguments{
\item{n}{a number, a sample size of first sample related to mu1}

\item{m}{a number, a sample size of second sample related to mu2}

\item{Delta}{a number, Delta = mu1 - mu2 is the
argument for the power function.}

\item{sig}{a number, common standard deviation of the 2 populations}

\item{alpha}{a number, the significance level of the test}
}
\value{
a number, gamma the power of the t-test to dectect delta
}
\description{
computes power function for two-sample t-testfor the
alternative mu1 > mu2. This test is discussed in Example 4.6.2,
page 278. Power functions for two-sample t-tests are discussed in
Example 8.3.3 on page 493.
}
\examples{
n <- 20
m <- 15
Delta <- 5
sig <- 10
alpha <- 0.05
tpower2g2(n, m, Delta, sig, alpha)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
