% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpowerg.r
\name{tpowerg}
\alias{tpowerg}
\title{Power of the One Sample t-Test}
\usage{
tpowerg(mu0, sig, n, alpha = 0.05, byv = 0.1)
}
\arguments{
\item{mu0}{Null population mean.}

\item{sig}{Population standard deviation.}

\item{n}{Sample size.}

\item{alpha}{Level of significance.}

\item{byv}{Increment in sequence for creating mu1.}
}
\value{
Vector of powers from power fucntion.
}
\description{
Computes a graph of the power curve of a one sample t-test
of mu=mu0 vs mu!=mu0 as shown in Example 8.3.2 on page 492 of the book.
This function assumes that X_1,...,X_n is a random sample on X that has a
N(mu, sigma^2) distribution. We are testing H_0: mu=mu_0 versus
H_1: mu != mu_0, where mu_0 is specified. Thus the likelihood ratio test
statistic is

\eqn{t(X_1,...,X_n)}

\eqn{= \sqrt(n)(Xbar - \mu_0)/
\sqrt((\sum(1 -> n)((X_i - Xbar)^2)/(n - 1)))}

\eqn{= (\sqrt(n)(Xbar - \mu_0)/\sigma) / (\sqrt( \sum(1 -> n)(X_1 - Xbar)^2 /
\(\sigma^2(n-1))))}

The hypothesis H_0 is rejected at level \eqn{\alpha} if
\eqn{|t| \ge t_\alpha / (2,n-1)}.

See page 492 and Exercise 8.3.5 on page 497 of the book.
}
\examples{
tpowerg(50, 10, 25)
tpowerg(50, 10, 25, 0.04, 0.2)

}
\references{
Hogg, R. McKean, J. Craig, A (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson
}
