% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpowerg2.r
\name{tpowerg2}
\alias{tpowerg2}
\title{Graphs a Two Sample T-Test}
\usage{
tpowerg2(mu0, sig, n, alpha = 0.05, byv = 0.1)
}
\arguments{
\item{mu0}{a number, null population mean null value}

\item{sig}{a number, population standard deviation usually greater than 1}

\item{n}{an integer, sample size}

\item{alpha}{a number, the signifigance level, defaults to 0.05}

\item{byv}{a number, increment in sequence for mu 1, defaults to 0.1}
}
\value{
gammas vector of powers from power function
}
\description{
Graphs the power of the one-sample, two-sided $t$-test
for the mean. The hypothese are:\cr
H0: mu = mu0 vs. H1: mu != mu0.\cr
See Example 8.3.2 on page 492 for discussion of this test andits power.
Exercise 8.3.5 continues the discussion.
}
\examples{
mu0 <- 50
sig <- 10
n <- 25
alpha <- .04
byv <- .15
tpowerg2(mu0, sig, n, alpha, byv)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
