% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttable.r
\name{ttable}
\alias{ttable}
\title{t-Distribution Table}
\usage{
ttable()
}
\value{
\code{tab} is the t-distribution table.
}
\description{
Generates the t-distribution probabilities for the selected quantiles
and degrees of freedom.
}
\details{
This table more specifically produces the t-distribution table
for the \emph{0.900}, \emph{0.950}, \emph{0.975}, \emph{0.990}, \emph{0.995},
\emph{0.999} quantiles. The initial row indicates the selected quantiles and
the inital column indicates 'r' degrees of freedom See page 710 for more details.
}
\examples{
ttable()

}
\references{
Hogg, R., McKean, J., Craig, A. (2018) Introduction to Mathematical
            Statistics, 8th Ed. Boston: Pearson.
}
