% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldpois.r
\name{waldpois}
\alias{waldpois}
\title{Wald Test Statistic}
\usage{
waldpois(x, theta0)
}
\arguments{
\item{x}{a vector of numeric elements, the sample.}

\item{theta0}{a vector of numeric elements, each element is a mean you are checking
the set against.}
}
\value{
this function returns a floating point number for each element, the test
statistic for each element in theta0
}
\description{
Computes the Wald test statistic for the null hypothesis
H0: theta = theta_0 where theta is mean of a Poisson distribution.
This is discussed in Exercise 6.3.17 on page 386 of HMC.
}
\details{
this function uses the following mathematical function:
(sqrt(length(x) / mean(x)) * (mean(x) - theta0))^2
}
\examples{
x <- c(3.5, -6, 5, 4.444, 10)
theta <- 23
waldpois(x, theta)

}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
