% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{waterwheel}
\alias{waterwheel}
\title{waterwheel}
\format{A list containing 20 rows and 2 variables
\describe{
  \item{numbrev}{something}
  \item{indgrp2}{not sure}
}}
\usage{
waterwheel
}
\description{
A data set containg 20 observations about waterwheel
}
\keyword{datasets}
