% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wil2powsim.r
\name{wil2powsim}
\alias{wil2powsim}
\title{Wilcox Power Simulation}
\usage{
wil2powsim(n1, n2, nsims, eps, vc, Delta = 0, alpha = 0.05)
}
\arguments{
\item{n1}{Sample size 1.}

\item{n2}{Sample size 2. This value must be divisible by the number of
elements in vector Delta.}

\item{nsims}{Number of iterations (simulations).}

\item{eps}{Contamination rate (epsilon).}

\item{vc}{Standard deviation of contaminated part.}

\item{Delta}{Vector of shifts in location between models. Sample size 2 (n2)
must be divisible by the number of ekements in this vector.}

\item{alpha}{Level of significance of the test.}
}
\value{
Vector containing empirical power of MWW test and t-test.
}
\description{
Iterates over specified number of simulations. At each step,
generate independent samples, compute each test statistic, and record whether
or not each test rejected. For each test, its empirical power is its number
of rejections divided by the number of simulations.
See section 10.4.4 on page 605 of the book.
}
\examples{
# Example where variables are initialized
# then passed into function.
n1 <- 30
n2 <- 30
nsims <- 10000
eps <- 0.2
vc <- 10
Delta <- c(-3, 3, 1)
alpha <- 0.25
results <- wil2powsim(n1, n2, nsims, eps, vc, Delta, alpha)


# Example where values are passed directly
# into function, along with a default param
# override for the Delta param.
results <- wil2powsim(30, 30, 10000, 0.20, 10, c(-3, 3, 1), 0.25)

}
\references{
Hogg, R. McKean, J. Craig, A (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson
}
