% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{aztdoses}
\alias{aztdoses}
\title{aztdoses}
\format{A list with 20 rows and 2 variables:
\describe{
  \item{azt}{no clue what this means}
  \item{dose}{maybe has something to do with how much given}
}}
\usage{
aztdoses
}
\description{
A data set containg 20 azt and doses
}
\keyword{datasets}
