% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bb}
\alias{bb}
\title{bb}
\format{A list with 59 rows and 6 variables:
\describe{
  \item{hand}{right or left}
  \item{height}{height in inches}
  \item{weight}{weight in pounds}
  \item{hitind}{something to do with hit?}
  \item{hitpitind}{something}
  \item{average}{some kind of an average}
}}
\usage{
bb
}
\description{
A data set containing 59 with multiple types of data
}
\keyword{datasets}
