% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{beta30}
\alias{beta30}
\title{beta30}
\format{A list of 30 rows in 1 variable
\describe{
  \item{x}{a beta}
}}
\usage{
beta30
}
\description{
A data set containg 30 betas
}
\keyword{datasets}
