% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{braindata}
\alias{braindata}
\title{brain data}
\format{A list of 40 rows and 7 variables
\describe{
  \item{gender}{gender of the person male or female}
  \item{FSIQ}{something}
  \item{VIQ}{something}
  \item{PIQ}{something}
  \item{weight}{weight of the person in pounds}
  \item{height}{height of the person in inches}
  \item{MRI_Count}{something}
}}
\usage{
braindata
}
\description{
A data set containg 40 observations about human brains
}
\keyword{datasets}
