% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{conductivity}
\alias{conductivity}
\title{conductivity}
\format{A list of 18 rows with 3 variables
\describe{
  \item{Conductivity}{the value of conductivity}
  \item{Binder}{something}
  \item{Aggregate}{something}
}}
\usage{
conductivity
}
\description{
A data set containg 18 observations about conductivity
}
\keyword{datasets}
