% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{earthmoon}
\alias{earthmoon}
\title{earthmoon}
\format{A list containg 7 rows and 1 variable
\describe{
  \item{ratios}{earth to moon ratios}
}}
\usage{
earthmoon
}
\description{
A data set containing 7 ratios
}
\keyword{datasets}
