% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{eg4.4.4data}
\alias{eg4.4.4data}
\title{eg 4.4.4 data}
\format{A list with 15 rows and 1 variable
\describe{
  \item{x}{some variable from 4.4.4}
}}
\usage{
eg4.4.4data
}
\description{
A data set containing 15 values
}
\keyword{datasets}
