% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex6111}
\alias{ex6111}
\title{Ex 6.1.1 Data}
\format{A list containing 25 rows and 1 variable
\describe{
\item{x}{some variable from 6.1.1}
}}
\usage{
ex6111
}
\description{
A data set containg 25 values
}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\keyword{datasets}
