% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex763data}
\alias{ex763data}
\title{ex 7.6.3 data}
\format{A list containing 30 rows and 1 variable
\describe{
  \item{x}{some variable from 7.6.3}
}}
\usage{
ex763data
}
\description{
A data set containg 30 values
}
\keyword{datasets}
