% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{examp1053}
\alias{examp1053}
\title{Ex. 10.5.3 Data}
\format{A list containing 15 rows and 2 variables
\describe{
\item{x}{some variable from 1053}
\item{y}{some variable frim 1053}
}}
\usage{
examp1053
}
\description{
A data set containg 15 x and y values
}
\details{
see page 615 of the book
}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\keyword{datasets}
