% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{exercise8316}
\alias{exercise8316}
\title{exercise 8.3.16 data}
\format{A list containing 15 rows and 2 variables
\describe{
  \item{x}{some variable from 8.3.16}
  \item{y}{some variable from 8.3.16}
}}
\usage{
exercise8316
}
\description{
A data set containg 15 x and y values
}
\keyword{datasets}
