% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lifetimemotor}
\alias{lifetimemotor}
\title{life time motor}
\format{A list containing 20 rows and 1 variable
\describe{
  \item{lifetime}{life of a motor}
}}
\usage{
lifetimemotor
}
\description{
A data set containg 20 observations about motor life time
}
\keyword{datasets}
