% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{men1500m}
\alias{men1500m}
\title{men 1500m}
\format{A list containing 27 rows and 2 variables
\describe{
  \item{year}{the year this took place}
  \item{time}{the time taken to run in minutes}
}}
\usage{
men1500m
}
\description{
A data set containg 27 observations about mens 1500m runs
}
\keyword{datasets}
