% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mix668}
\alias{mix668}
\title{mix668}
\format{A list containing 20 rows and 1 variable
\describe{
  \item{x}{some variable from 6.6.8}
}}
\usage{
mix668
}
\description{
A data set containg 20 observations of x values
}
\keyword{datasets}
