% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{olym1500mara}
\alias{olym1500mara}
\title{olympic 1500 marathon}
\format{A list containing 20 rows and 3 variables
\describe{
\item{year}{the year it took place}
\item{m1500}{the winning time in seconds of the 1500 meter run}
\item{marathon}{the winning time in seconds of the marathon}
}}
\usage{
olym1500mara
}
\description{
A data set containg 20 observations about olympic 1500 meter run and the marathon
}
\details{
see pages 545 and 633 of the book.
}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\keyword{datasets}
