% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesampsgn.R
\name{onesampsgn}
\alias{onesampsgn}
\title{One-Sample Sign Analysis}
\usage{
onesampsgn(x, test = FALSE, alt = 0, theta0 = 0, alpha = 0.05,
  maktable = TRUE, plotb = FALSE)
}
\description{
Calculates a distribution free confidence interval for a sample
median by deriving for the 100pth distribution percentile of a continuous random
variable.
}
