% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisrand.R
\name{poisrand}
\alias{poisrand}
\title{Simulating Poisson Processes}
\usage{
poisrand(n, lambda)
}
\arguments{
\item{n}{Iterations to repeat (number of simulations).}

\item{lambda}{Mean variable in function.}
}
\value{
Vector of realizations from poisson distribution.
}
\description{
Generate n simulations of a Poisson Distribution. See Example
4.8.2 on page 295 of 8th edition.
}
\examples{
# Example where parameters are passed in as variables.
n <- 1000
lambda <- 5
result <- poisrand(n, lambda)

# Example where parameters are passed in as values.
result <- poisrand(1000, 5)

}
