% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{punter}
\alias{punter}
\title{Distance of Punts}
\format{A list containing 13 rows and 7 variables
\describe{
\item{distance}{the average distance in feet of 10 punts}
\item{strength}{the strength}
\item{hangtime}{the time the ball was in the air}
\item{lf}{a predictor}
\item{llstrength}{a predictor of strength}
\item{rf}{a predictor}
\item{rl}{a predictor}
}}
\usage{
punter
}
\description{
A data set containg 13 observations about Punters
}
\details{
See page 360 of the book.
}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\keyword{datasets}
