% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{quailldl}
\alias{quailldl}
\title{quailldl}
\format{A list containing 39 rows and 2 variables
\describe{
  \item{treat}{i dont know}
  \item{ldl}{not sure}
}}
\usage{
quailldl
}
\description{
A data set containg 39 observations about quailldl
}
\keyword{datasets}
