% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{regr1}
\alias{regr1}
\title{regr1}
\format{A list containing 15 rows and 2 variables
\describe{
  \item{actscores}{scores gotten on act}
  \item{examscores}{scores gotten on exams}
}}
\usage{
regr1
}
\description{
A data set containg 15 observations about test scores
}
\keyword{datasets}
