% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sec951}
\alias{sec951}
\title{sec 9.5.1}
\format{A list containing 24 rows and 3 variables
\describe{
\item{x}{some variable x}
\item{fa}{some other variable fa}
\item{fb}{soem other variable fb}
}}
\usage{
sec951
}
\description{
A data set containg 24 observations from section 9.5.1
}
\details{
See page 539 of the book.
}
\references{
Hogg, R. McKean, J. Craig, A. (2018) Introduction to
Mathematical Statistics, 8th Ed. Boston: Pearson.
}
\keyword{datasets}
