% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{shoshoni}
\alias{shoshoni}
\title{shoshoni}
\format{A list containing 20 rows and 1 variable
\describe{
  \item{shoshoni}{something}
}}
\usage{
shoshoni
}
\description{
A data set containg 20 observations about shoshoni
}
\keyword{datasets}
