% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{speedlight}
\alias{speedlight}
\title{speedlight}
\format{A list containing 66 rows and 1 variable
\describe{
  \item{speed}{the speed in some units}
}}
\usage{
speedlight
}
\description{
A data set containg 66 observations of speeds
}
\keyword{datasets}
