% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sulfurdio}
\alias{sulfurdio}
\title{sulfur dioxide}
\format{A list containing 24 rows and 1 variable
\describe{
  \item{sulfurdioxide}{the level of sulfur dioxide}
}}
\usage{
sulfurdio
}
\description{
A data set containg 24 observations of sulfur dioxide levels
}
\keyword{datasets}
