% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{telephone}
\alias{telephone}
\title{telephone}
\format{A list containing 24 rows and 2 variables
\describe{
  \item{year}{the year of the observation}
  \item{numcall}{the average number of calls}
}}
\usage{
telephone
}
\description{
A data set containg 24 observations about phone calls
}
\keyword{datasets}
