% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tempbygender}
\alias{tempbygender}
\title{temp by gender}
\format{A list containing 130 rows and 2 variables
\describe{
  \item{temp}{something}
  \item{gender}{gender of the person male or female}
}}
\usage{
tempbygender
}
\description{
A data set containg 130 observations of temp by gender
}
\keyword{datasets}
